requirements.txt
asgiref==3.8.1
autopep8==2.3.1
Django==5.1.1
django-silk==5.2.0
djangorestframework==3.15.2
gprof2dot==2024.6.6
pycodestyle==2.12.1
sqlparse==0.5.1
tomli==2.0.1
typing_extensions==4.12.2
